replace the bespoke `OpenAiChatCompletions` adapter with `@effect/ai-openai-compat`
